fix skipping for debugging and use 2x 98-percentile of fragment length
when debugging with small reads, we'd skipp all reads and get a median
and 98th percentile of fragment length of 0. This simply saves skipped
reads and uses them if no other reads were found.

Also, adjust to use 2 * max_dist (which is 98th percentile) because
we can have reads a fragment length away on either side of the event.

we could instead use a much larger number for mediani (e.g. 60), but we
can test this later once we have the simpler details worked out.